GCF/crawl_and_translate_news/requirements.txt
functions-framework==3.*
requests >= 2.22.0
beautifulsoup4 >= 4.11.1
googletrans==3.1.0a0
pandas >=2.0.3
google-cloud-storage >= 2.10.0